--- .env
@@ -1,8 +1,0 @@
-DATABASE_URL=‘postgres://iypeenbt:tQywvRWSM6qQppSVmJoOGfhhyDAvTjP-@mahmud.db.elephantsql.com/iypeenbt’
-GITHUB_ID=0c881b48b3e0c7d1a0ff
-GITHUB_SECRET=8b7720df621bd662c473f0dc488b03e852ddba80
-GOOGLE_ID=74460954164-2bdjqbemvfdotpagulvbru9nmeclh5ep.apps.googleusercontent.com
-GOOGLE_SECRET=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
-NEXTAUTH_SECRET="supersecret"
-NEXTAUTH_URL="http://localhost:3000"
-
